fix(api): fixed modes for version 1.43.0
