+++ packages/api/test/modes/readModeRangeSlots.spec.ts
@@ -1,6 +1,6 @@
-import mockMsp from "./mockMsp";
-import { readModeRangeSlots } from "../src";
-import codes from "../src/codes";
+import mockMsp from "../mockMsp";
+import { Modes, readModeRangeSlots } from "../../src";
+import codes from "../../src/codes";
 
 beforeEach(() => {
   mockMsp.resetResponses();
@@ -125,151 +125,152 @@ describe("readModeRangeSlots", () => {
     mockMsp.setResponseForCode(
       [
         0,
-        3,
-        16,
+        0,
         32,
-        0,
-        3,
-        16,
-        34,
-        0,
-        5,
+        48,
         2,
-        18,
-        0,
-        0,
-        16,
-        32,
-        0,
-        0,
-        16,
-        32,
-        0,
-        0,
-        16,
-        32,
-        0,
-        0,
-        16,
-        32,
-        0,
-        0,
-        16,
-        32,
-        0,
-        0,
-        16,
-        32,
-        0,
-        0,
-        16,
-        32,
-        1,
-        0,
-        16,
-        32,
-        1,
-        0,
-        16,
-        32,
-        1,
-        255,
-        16,
-        32,
-        1,
-        255,
-        16,
-        32,
-        1,
-        255,
-        16,
-        32,
         2,
+        32,
+        48,
+        13,
+        1,
+        32,
+        48,
+        35,
+        1,
+        32,
+        48,
         0,
-        16,
-        32,
-        2,
-        255,
-        16,
-        32,
-        2,
-        255,
-        16,
-        32,
-        2,
-        255,
-        16,
-        32,
-        6,
         0,
-        16,
-        32,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
       ],
       codes.MSP_MODE_RANGES
     );
 
     mockMsp.setResponseForCode(
       [
-        0,
-        1,
-        1,
-        0,
-        1,
-        1,
-        0,
-        1,
-        1,
-        0,
-        1,
+        20,
         0,
         0,
-        1,
-        1,
         0,
-        1,
-        1,
-        0,
-        1,
-        1,
-        0,
-        1,
-        1,
-        0,
-        1,
-        1,
-        0,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
         2,
-        1,
-        1,
-        2,
-        1,
-        1,
-        2,
-        1,
-        2,
-        2,
-        1,
-        1,
-        6,
-        1,
-        1,
+        0,
+        0,
+        13,
+        0,
+        0,
+        35,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
+        0,
       ],
       codes.MSP_MODE_RANGES_EXTRA
     );
@@ -277,143 +278,143 @@ describe("readModeRangeSlots", () => {
     expect(await readModeRangeSlots("/dev/someport")).toEqual([
       {
         modeId: 0,
-        auxChannel: 3,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
-      },
-      {
-        modeId: 0,
-        auxChannel: 3,
-        range: { start: 1300, end: 1750 },
-        linkedTo: 1,
-        modeLogic: 1,
-      },
-      {
-        modeId: 0,
-        auxChannel: 5,
-        range: { start: 950, end: 1350 },
-        linkedTo: 1,
-        modeLogic: 1,
-      },
-      {
-        modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
+        range: { start: 1700, end: 2100 },
+        modeLogic: 0,
         linkedTo: 0,
-        modeLogic: 1,
+      },
+      {
+        modeId: Modes.HORIZON,
+        auxChannel: 2,
+        range: { start: 1700, end: 2100 },
+        modeLogic: 0,
+        linkedTo: 0,
+      },
+      {
+        modeId: Modes.BEEPER_ON,
+        auxChannel: 1,
+        range: { start: 1700, end: 2100 },
+        modeLogic: 0,
+        linkedTo: 0,
+      },
+      {
+        modeId: Modes.FLIP_OVER_AFTER_CRASH,
+        auxChannel: 1,
+        range: { start: 1700, end: 2100 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
         modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
         modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
         modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
         modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
         modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
         modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
-        modeId: 1,
+        modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
-        modeId: 1,
+        modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
-        modeId: 1,
-        auxChannel: 255,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
-      },
-      {
-        modeId: 1,
-        auxChannel: 255,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
-      },
-      {
-        modeId: 1,
-        auxChannel: 255,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
-      },
-      {
-        modeId: 2,
+        modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
       {
-        modeId: 2,
-        auxChannel: 255,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
-      },
-      {
-        modeId: 2,
-        auxChannel: 255,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 2,
-        modeLogic: 1,
-      },
-      {
-        modeId: 2,
-        auxChannel: 255,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
-      },
-      {
-        modeId: 6,
+        modeId: 0,
         auxChannel: 0,
-        range: { start: 1300, end: 1700 },
-        linkedTo: 1,
-        modeLogic: 1,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
+      },
+      {
+        modeId: 0,
+        auxChannel: 0,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
+      },
+      {
+        modeId: 0,
+        auxChannel: 0,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
+      },
+      {
+        modeId: 0,
+        auxChannel: 0,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
+      },
+      {
+        modeId: 0,
+        auxChannel: 0,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
+      },
+      {
+        modeId: 0,
+        auxChannel: 0,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
+      },
+      {
+        modeId: 0,
+        auxChannel: 0,
+        range: { start: 900, end: 900 },
+        modeLogic: 0,
+        linkedTo: 0,
       },
     ]);
